Add support for infinitely nested repeat blocks
The getTickForLine function now supports infinitely nested loops. It
makes use of recursion by having a function that evaluates the ticks
passing for one loop, and if that function finds a line starting with
"end" it will call itself again for that loop and add the value that's
evaluated from that function call to the ticks for the outer loop.