Change language server implementation
This moves the old language server implementation to
server/src/_tas-script and starts a new parser implementation in
server/src/tas-script.

The new implementation now works as follows:
    - Generate tokens for each line, containing information about the
      line, start and end indices, a type and the text contained
    - Analyse the tokens and remove all single and multiline comments,
      leaving the tokens that are actually relevant. Since each token
      contains precise position info that can be used directly in
      diagnostic creation, we can just remove these tokens without
      worring.
    - "Parse" these tokens. This means that we enforce a certain
      structure of tokens, while also extracting information like tick
      info, whether it is relative, and in the future tool info as well.

This commit is an intermidiary commit, as it removes a bunch of stuff to
later be re-implemented. The parser currently only validates version,
start, and framebulks (excluding tools).