Add simple hover provider
This re-implements the hover provider in a rather naive way:
The hover provider now just searches through all tokens in a line and
tries to find the token with the cursor in it. If it is a number token
and the next token is a '>', it will return the tick. Otherwise it will
go through all tool names and their arguments, trying to find one that
matches the text. If it finds one, it displays the corresponding
description.