Rewrite language server parser to use a tokenizer (#30)
* Change language server implementation

This moves the old language server implementation to
server/src/_tas-script and starts a new parser implementation in
server/src/tas-script.

The new implementation now works as follows:
    - Generate tokens for each line, containing information about the
      line, start and end indices, a type and the text contained
    - Analyse the tokens and remove all single and multiline comments,
      leaving the tokens that are actually relevant. Since each token
      contains precise position info that can be used directly in
      diagnostic creation, we can just remove these tokens without
      worring.
    - "Parse" these tokens. This means that we enforce a certain
      structure of tokens, while also extracting information like tick
      info, whether it is relative, and in the future tool info as well.

This commit is an intermidiary commit, as it removes a bunch of stuff to
later be re-implemented. The parser currently only validates version,
start, and framebulks (excluding tools).

* Re-implement tool parsing into the new language server implementation

The new language server implementation is now also capable of parsing
tools. This brought a change to the way we store the tool information
with it too, meaning we can better enforce optional arguments and
required arguments like for example setang expects.

* Add simple hover provider

This re-implements the hover provider in a rather naive way:
The hover provider now just searches through all tokens in a line and
tries to find the token with the cursor in it. If it is a number token
and the next token is a '>', it will return the tick. Otherwise it will
go through all tool names and their arguments, trying to find one that
matches the text. If it finds one, it displays the corresponding
description.

* Re-implement repeat support

* Re-implement active tools display

* Re-implement tick type toggle request (p2tas/toggleLineTickType)

* Re-implement tool & tool argument completion

* Re-implement completion of other keywords

This re-implements completion of other keywords, meaning e.g. "repeat",
"start" (and it's arguments) and "end", into the new language server
implementation.

* Remove old (commented-out) completion

* Remove old language server implementation

* Remove old language server implementation

* Add check tool

* Improve hover documentation provider

* Make certain units optional + tokenizer improvements

* Add error if first framebulk is relative

* Add error for missing lines

* Trigger completion on '|'

* Correctly add active tools with determined order

* Don't subtract the current line's tick from new tools from the line

* Allow numbers followed by strings after map / save start