Add support for "autoaim ent <name>"
This adds support for the conditional parameter "ent" added to autoaim
in SAR. In this commit, the way these conditional arguments are parsed
has also been changed. It now works by specifying children for an
argument, which are parsed immediately after this argument. In addition,
you can also specify children that are parsed if the argument didn't
match (otherwiseChildren).
fix #42