Allow variable number of tokens in start statement
With the start types "map", "save" and "cm", a variable number of tokens
are now allowed, as long as they are adjacent. This allows for all
characters to be inputted, regardless of how they are tokenized.

Examples:
"1/2/3" - Parsed as 5 separate tokens, but all are adjacent -> no error
"1/2 3" - First 3 tokens accepted (^), 4th not adjacent -> error on
fourth

fix #46