Don't highlight tools and tool arguments outside of commands field
Now, when the tool names / their arguments appear outside of the
commands field, they won't be colored. This makes it clear to the user
that this is the wrong field for the tool, while not being intrusive.
It is achieved with a positive lookbehind, matching for four '|'s or two
'>'s with any other character around them.

fix #34