Improve handling of path arguments
The `start` and `rngmanip` statements both (might) take a path argument,
which needs to be handled in a special way due to how the script is
tokenized. We now have a unified method in the parser for accepting such
strings.

The method `acceptConsecutiveTokens()` will accept any number of tokens
that are adjacent to each other (i.e. the first token's end is the same
as the next token's start, due to the end being exclusive) and return
how many tokens were accepted. This makes it possible to accept
continuous strings (like e.g. paths), but reject anything separated by a
space.